ai-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
pydantic-settings>=2.0
python-multipart>=0.0.9

# AI / Audio
# NOTE: torch & torchaudio are installed as CPU-only wheels by the Dockerfile
#       (via https://download.pytorch.org/whl/cpu) BEFORE this file is read,
#       so pip treats them as already satisfied and skips the CUDA build.
#       For local dev without Docker, run:
#         pip install torch torchaudio --index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
torchaudio>=2.0.0
speechbrain>=1.0.0
numpy>=1.24.0
scipy>=1.11.0
resemblyzer>=0.1.3
demucs>=4.0.1
audio-separator[cpu]>=0.44.1
noisereduce>=3.0.0
pyannote.audio>=3.1.0

# Database
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29.0
pgvector>=0.3.0

# Object storage
minio>=7.2.0

# Worker
dramatiq[redis]>=1.16
redis>=5.0

# Observability
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.44b0
opentelemetry-exporter-otlp-proto-http>=1.20.0
opentelemetry-instrumentation-sqlalchemy>=0.44b0
opentelemetry-instrumentation-logging>=0.44b0
